+++ order-service/src/main/java/com/ed/orderservice/presentaion/web/response/domain/orderCreate/NewPurchaseOrderMessageResponse.java
@@ -1,4 +1,4 @@
-package com.ed.orderservice.presentaion.web.response;
+package com.ed.orderservice.presentaion.web.response.domain.orderCreate;
 
 import com.ed.orderservice.application.port.in.dto.OrderItemDto;
 import com.ed.orderservice.domain.vo.order.Order;
